MLabel: Use several font bindings instead of a single one
This allows for partial override of label properties without breaking
the internal binding.

For instance, in the situation where the user wants a standard label
but with a different text color, if he would do Label { font.color: "blue" }
he would be breaking the property binding internal to the Label component.

Now since we have independent bindings, in the above situation all bindings
would keep alive, except for the color one of course.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>