MeegoTouch Pages: added a PageContainer for each Page
The PageContainer encapsulates each Page element. This enable us to draw
the background on the whole screen size (as meegotouch does), as the
PageContainer has the same size of the Window.

Now in Pagemanager we keep a list of PageContainers, which contain a
'page' property for accessing their respective page. Also the sliding
logic moved from Page to PageContainer.

As it is now, Page is the interface most of the application developer
will use, while PageContainer can contain more platform specific
behavior (could be a private type, as it is currently instantiated internally
in Window).

Reviewed-by: Frederik Gladhorn