MeeGo Touch: add new components for Actions
This commit adds new components to provide a feature similar the
libmeegotouch's MAction and the toolbar visual element. Our Action
component is a QObject representing an action with a 'triggered' signal.

An action also has an 'item' property that contains the visual component
that will represent the action in the toolbar. Usually it's not necessary
to set it, but set the iconId instead. This will use the default item for
actions which is the ActionButton.

Next commits will make the Actions be usable by putting the 'item' element
in the toolbar.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>