MeeGo Touch: add toolbar support with Actions
Adds a toolbar for both landscape and portrait modes, following the
behaviour of libmeegotouch. To fill the toolbar, you declare a list of
Actions inside the 'actions' property of a Window or a Page.

The toolbar will be filled with the 'items' that represent each
Action. The toolbar will always show Window's actions and the current
Page's actions, except when their property 'interactive' is set to
false.

There isn't support for changing the action lists in "runtime",
however the 'interactive' property can help in some situations. Also
not done is the support for text in the default ActionButtons.

This commit also moved the TitleMenuButton to inside the titlebar and
added a comment on why the HomeButton wasn't inside the titlebar: it
needs to stand out on top of the query dialog (title menu doesn't
need, so keep inside titlebar).

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>