MStyleWrapper: Fix logic to update resources when orientation changes
The previous logic in MStyleWrapper was mimicking what MStyleContainers
do, that is to watch MApplication::currentWindow()->orientation().

In our enviorment however we never create MWindows, thus there is never
a currentWindow() we can look into.

The new approach is to get the current orientation from MInputMethodState
that is changed by MDeclarativeScreen::setOrientation().

As a note, that's an additional reason for accessing MThemes directly
rather than using MStyleContaneirs, of which we don't have control.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Leonardo Sobral Cunha <leo.cunha@nokia.com>