Slider now implemented.
This change adds the slider to the gallery. While developing the
slider, I found that I had to rewrite the range model quite a bit.
The reason is that you never have any guarantees in which order
QML will assing values to minimum, maximum, value etc on the
slider. This means that if the model e.g. assumes that it has a non-empty
value range when a value is set, it might fail (a range of zero will make
any mapping between value and position end up as zero). Besides,
because of the way QML binds values to properties, a value
that lies outside the valid value range at one point (and therefore
will be bound to the edge of the range), might end up valid later on.
So the model cannot force a value inside the range when setting
it. Moreover, even when changing the range of the model, we need to check if
that caused a change of 'effective' value.