Meego Switch: Fix reactive margins and background color
The property that was being used to enlarge the switch click area
was the wrong one. In MeegoTouch widgets use "reactiveMargin" properties
to define whether they will react to mouse events outside their area.

In other words, MeegoWidgets have a QGraphicsItem::shape() that is larger
than their boundingRect by the amount defined in style()->reactiveMargin*.

This commit fixes that and uses a simpler approach which is to grow the
MouseArea outside the component boundaries, which is what happens in MeegoTouch
as well.

 -  -  -  -  -  -  -  -  -  - - <-------- Area sensible to mouse events
 .    ___________________     .
 .   /                   \ <------------- Item boundary, used as the reference
 .   \___________________/    .           to other items that anchor to it, etc.
 .                            .
 -  -  -  -  -  -  -  -  -  - -

Also fixing the painting of the component background. The logic was changed
in the latter MeegoTouch versions and we had to copy it again.
The current solution is not completly equal to what the original version does
but gets close and does not require changes to the C++ primitive.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>