MeegoTouch ListItem: moved list items to src/Meego
Moved ListItem and BasicListItem to MeegoTouch items.
Added a mouse area to ListItem and a clicked signal.

Also removed the rounded corners from the first and last
list item's background (rounded corners are only used for
grouped list items). Exposed the tilePosition as a
property of ListItem (the default is verticalCenter).