Class hierarchy change
Let MDeclarativePrimitive derive from MDeclarativeImplicitSizeItem. The
reason for this is that also primitives makes use of implicit size (e.g.
MScalableImage sets it to the unscaled size of the image). And this
information is useful to query from QML in many cases.

The second part of the patch makes use of this structural change to
remove warnings about drawing images inside empty rects. It should be
acceptable to set the width of e.g. MScalableImage to zero, as it can
easily happend if you bind the width to something else.