MeeGo Touch: allow specify parent class name in StyleWrapper
This is the simplest interface possible for specifying the style
parent class for our MStyleWrapper components. The function MTheme::style()
now supports a list of string lists

  - [parentClass, superClass-of-parentClass, superSuperClass-of-parentClass...]
  - [grandparentClass, superClass-of-grandparentClass, ...]
  - ...

But, since the themes we care about right now still don't depend on
the full hierarchy for getting the right theme, we can live by
specifying just the correct class name of parent.

Support for passing more data (like the class names of all ascendants)
or properly filling each list with the superclasses can be added
later without problems.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>