MeeGo Touch: demonstrate alternate spinner style
Export a property '__styleParentClass' and use it in the test
program. For the Spinner, the special styles appear inside other
widgets, so they'll be detail of implementation of those widgets. We
might want to add a better API later.

As a bonus, get rid of the "poor's man slider" now that we have a real
one :-)

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>