Themebridge primitives: Use synchronous MTheme::pixmap() method
The method we use to fetch style resources from MTheme used to
be asynchronous, what required us to do some painful bookkeeping
of which resources were available yet and which were not.

This behavior has changed in the latest libmeegotouch versions,
this commit updates our primitives to take advantage of that.

Code is pretty much simplified, there's no need to "checkPendingPixmap"
anymore and any logic that used to be done there, can now be done
earlier, in "fetchStyleData()".

If you experience pixmap / drawing problems after this commit,
ensure you have an updated version of libmeegotouch.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>