MDeclarativeIcon: Use item size to render pixmaps accordingly
This primitive was always asking MTheme for pixmaps of a standard
size. This is OK when we don't have a defined item size, but when
we have, we should tell MTheme about our size so that internally
it can render SVGs to the appropriate size.

An immediate result of this commit is that the "Close" button on
the top-right edge of the screen is no longer blurry.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>