MeeGo Touch: don't verify style class in StyleWrapper
Since libmeegotouch revision c6edf89223623494d49c1b97b2fd7e9db6870a79,
the MClassFactory is not public anymore, so we can't verify whether the
style class exists or not.

We were doing this because MTheme::style() assumed that it should exist, so
we needed a way to fail safe. I proposed a merge request to make it verify
this assumption and return NULL.

    http://meego.gitorious.org/meegotouch/libmeegotouch/merge_requests/851

Note that, this commit is independent of the one in the merge request,
because we need to remove the usage of MClassFactory regardless of how
MTheme::style() improves.

The issue was spotted by Max Waterman, that proposed a patch
(thanks!). This is a modified version of the original patch.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>