Remove QButtonModel
The QButtonModel was an attempt of extracting the behaviour of a
"button" into one helper non-visual object, that would be exported to
QML. This code was initially based on QAbstractButton, and still had
some artifacts from that time.

In practice, for the MeeGo and Mx, we ended up using MouseArea and
separating the mutual exclusive bits into the Checkable element
(working together with CheckableGroup). See the practical results of that
in examples/checkable.

People working on the "custom widgets" branch are investigating other
ways to "package" the features of a button and allow full
customization. If you are interested see there. One difference is that
their approach include a MouseArea inside -- which might end up being
more practical.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>