Make tests work "out-of-the-box" in Unix systems
The issue here is that the tests links with a QML module that we
build. But this file isn't installed on the system in a typical library directory,
but inside the QML modules structure, so it is not usually visible.

Adding to rpath of the tests mean that the tests will look in the
compiled library that is available inside the src/module directory. I
think this is a good compromise. The variable I used is for Unix only, so
the solution might not cover the Windows case. :-(

There might be a better way to solve that, but I don't know qmake that well. :-P
So if you know how to solve this in a better way, feel free to commit a fix.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>