Replaced PositionIndicator with ScrollDecorator for MeeGo
Uses the latest API as of today. The following pattern is recommended
when using the ScrollingDecorator with a Flickable (or any of its
deriving classes)

Item {
    // Do your anchoring/geometry setting here

    Flickable {
        id: theFlickable
        anchors.fill: parent
    }

    ScrollDecorator {
        flickable: theFlickable
    }
}

Also, when both the horizontal and vertical indicators are showing,
they will no longer overlap in the bottom right corner.

Examples updated and PositionIndicator removed.

Reviewed-by: Leo Cunha