MeeGo Touch: new Slider implementation
Add a new SliderTemplate element, that implements the slider component
with the basic pieces loaded as Components. It has a dummy visuals
based on rectangles. This allows a separation between the style specific parts and the real
functionality parts.

It already follows the Common API being defined in
http://bugreports.qt.nokia.com/browse/QTCOMPONENTS-200 plus some extensions
for functionality not defined there.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
             Tomas Junnonen <tomas.junnonen@nokia.com>