Added support for Dialog, QueryDialog and ViewMenu
Dialogs have a title bar, a content field and a button row. They are
used to display messages, ask simple Yes/No- and more complex questions
or deliver input methods. They can be modal (other layers/windows are
blocked), but maybe not.

With this commit a dialog is created in the following way:

Dialog {
  id: myDialog
  modal: true
  title: qsTr("Dialog")
  content:
    Text {
      id: text
      font.pixelSize: 22
      color: "white"
      text: qsTr("Some message")
    }
  buttons: Row {
    anchors.horizontalCenter: parent.horizontalCenter
    Button {id: b1; text: "OK"; onClicked: myDialog.accepted()}
    Button {id: b2; text: "Cancel"; onClicked: myDialog.rejected()}
  }
}

The "title" property expects a value of type string, "content" and
"buttons" expect Items (so you can draw a Rectangle inside the button
row instead of, yeah, just a row with buttons). If the properties are
left empty, nothing is drawn at all.

The emitted signals are:

signal activated
signal deactivated
signal accepted
signal rejected

For displaying a dialog, just set its state to "visible". To dismiss, use
the signals "accepted" or "rejected" or set its state to "hidden".

The QueryDialog is used in the same way, but has the additional
properties "message", "acceptButtonText" and "rejectButtonText":

QueryDialog {
  ...
  message: qsTr( "Do you really want to quit?")
  acceptButtonText: qsTr("Of course")
  rejectButtonText: qsTr("Not sure")
  ...
}

A view menu contains MenuItems. It is constructed in the following way:

Menu {
  id: myMenu
  content:
    Column {
       width: myMenu.width
       MenuItem {id: b1; text: "Copy"; onClicked: myMenu.accept()}
       MenuItem {id: b2; text: "Cut"; onClicked: myMenu.accept()}
   ...
  }
}