QRangeModel: fix tests when qreal is float
During the creation of the tests data, we define some columns as
qreals but later create the rows passing number literals that the compiler
may understands as doubles even if qreal is defined as a float.

This will cause assertion in the test data internal mechanism, because
we are feeding the column with the wrong type.

The fix is to make sure that the numbers are all qreals. The same
solution appears in Qt tests as well.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Leo Cunha