MeeGo Label: Remove dependency on ThemeBridge
Following the latest trend, I'm removing the use of themebridge
in the Label component. Instead I'm defining hardcoded values.

As an implementation detail, a contants file was created. The idea
in near future is to completely remove the Label component.

The WidgetsGallery example was updated accordingly.

Components that were using the old styleObjectName property (feature
of themebridge) are now customizing the component themselves, by
overriding the relevant properties.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Thiago de Barros Lacerda <thiago.lacerda@openbossa.org>

Reviewed-by: Leonardo Sobral Cunha <leo.cunha@nokia.com>