MeeGo Touch: implement CheckBox without old themebridge
Implements the CheckBox using the new image provider. Note that this
follows the themeing structure for the non-development theme. The
development theme still needs to be updated by MeeGo Touch graphic
designer team.

Also added an extra parameter to the UIConstants, that takes defines
the "extra" border that an mouse area gets when pressed. Got this
value from mviewconstants.h.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Thiago de Barros Lacerda <thiago.lacerda@openbossa.org>