MeeGo Touch: make calculator work again
The calculator was outdated, relying on the old TextEdit. The TextEdit was used
as a readOnly place to put the text. The horizontal alignment was also used.

Since the alignment is not available in the current TextField, I took the approach
of getting the information I needed and build the interface using a BorderImage for
the background and copying the TextField font property for my Text that display the
digits.

This might not be the ideal solution, but I'm pushing because at least makes the
calculator work again, we can fix it properly later ;-)

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>