Handle the arguments inside our api check
Properly deal with args using QCoreApplication::arguments().
With this patch, the correct way of using the tool is:

In order to test a module you can do like below:
$ ./tst_apicheck -module "Qt.labs.components.custom 1.0"

You can also pass an import path in the second parameter:
$ ./tst_apicheck -module "Qt.labs.components.custom 1.0" -importpath /mywork/imports

It's important to note that for now it defaults to testing the MeeGo
components if the above arguments are not specified.

We also pass the arguments to the tests, so we are able to output using
xml or to a file for example (and all the options of a regular QTest).

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>
Reviewed-by: Anselmo Lacerda S. de Melo <anselmo.melo@openbossa.org>