Added missing QMAKE_POST_LINK steps for native imports
"Cross-platform" auto tests are going to be modified to import native
components. This ensures that the same tests can be run with both
meego and symbian3 styles.

The idea is to mimic the imports installation hierarchy within the
build tree. This is needed in order to be able to run tests locally
without relying on 'make install' - just like the CI system does.

Reviewed-by: pending