Improve and partially fix readOnly test on text area
Improve the test by testing the functions that are affected by this property
like cut/paste. Also, reset the readOnly property to false to ensure that it
will not make other tests fail.

However this test still needs to get fixed in order to make the text property
work properly when it's readOnly (right now it's a separate property and not
an alias to TextEdit.text property. This is due a bug that was fixed only in
Qt 4.7.2). As soon as we can rely on a newer version of Qt we should properly
fix the components.

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>
Reviewed-by: Anselmo Lacerda S. de Melo <anselmo.melo@openbossa.org>