Improve 'select' tests on text area
Make sure we use QMetaObject::invokeMethod instead of creating a declarative
expression and evaluating it, as it would be much slower than doing directly
and now the code seems a little bit clearer than it was.

On selectedText() test, make sure to test the different modes of selection:
select(int,int), selectAll() and selectWord().

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>
Reviewed-by: Anselmo Lacerda S. de Melo <anselmo.melo@openbossa.org>