Fix slider auto test
This commit changes from an API test to functionality test. API tests
will be done by the apicheck tool. There are 3 expected failures in
this test that are related to the way we use enums in QML (ints).

We can't limit the values to known enum values as we map them to ints
and because of that the test will fail. A new bug is open for this
task:

http://bugreports.qt.nokia.com/browse/QTCOMPONENTS-414

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>
Reviewed-by: Thiago de Barros Lacerda <thiago.lacerda@openbossa.org>