Fix validator test for text field
As text field saves the property as just a QValidator, we should treat it
just as a QValidator when getting it back from the text field. This way we
can test everything. Also we should reset input mask and other properties
when we start testing the validator, otherwise acceptableInput() starts
to have a crazy behaviour.

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>
Reviewed-by: Anselmo Lacerda S. de Melo <anselmo.melo@openbossa.org>