Change text area wrap mode's test
We can rely on the font being used to do the test anymore. Because of
this let's do the text just as QDeclarativeTextEdit and test the behaviour
of the component instead of testing wrapMode itself (we can let this test
to Qt's tests anyway).

Also, add the test for the range of options. It's an expected failure due to
the bug:

http://bugreports.qt.nokia.com/browse/QTCOMPONENTS-414

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>
Reviewed-by: Anselmo Lacerda S. de Melo <anselmo.melo@openbossa.org>