Symbian Button cleaned up.
Removed the unnecessary __style property and moved the
autoRepeatInterval property as internal. Item is replaced
by QtObject as it is lighter. QtObject is used to hide private data.
StateGroup is used to hide the internal states of the button.
Moreover, a property initialization that caused circularity
was removed. The unit test was changed accordingly.

Merge-request: 268
Reviewed-by: J-P Nurmi <j-p.nurmi@nokia.com>