Adding a way to validate QML defined properties
When user creates a component on a QML file, that is used on another
component, it must be checked differently than comparing it with a
known Qt component.
QML defined components always follow the pattern:
"<ComponentName>_QMLTYP_<Number>*". Due to that, a new method was
created, in order to see if the type name of the property contains a
substring that mathes the pattern.

Signed-off-by: Thiago de Barros Lacerda <thiago.lacerda@openbossa.org>
Reviewed-by: Anselmo Lacerda S. de Melo <anselmo.melo@openbossa.org>