Symbian: fix a graphics issue of Button and ToolButton in ButtonRow.
Button and Toolbutton made to fetch graphics based on the visible
siblings. ButtonGroup modified to correspond with the changes and
the code has been refactored a bit as well. Moved common logics of
Button and ToolButton to ButtonRow.Previously there was some
duplicate code both in Button and ToolButton that was meant for
dissecting which buttons are visible siblings. This was
transferred to ButtonRow and both components now make use of it.

Merge-request: 473
Reviewed-by: J-P Nurmi <j-p.nurmi@nokia.com>