Symbian: Fix slider showing the tooltip when focus moved out
Changed Slider to show the tooltip only on a key press event it consumes
instead of showing it on all the key press events it sees. This change
also impacts how the common API property pressed is set as a result of
key input. After this change it's only set as a result of a key pressed
event that the Slider will consume. Earlier it was set on all the key
pressed events.

Task-number: QTCOMPONENTS-609
Merge-request: 506
Reviewed-by: J-P Nurmi <j-p.nurmi@nokia.com>
Reviewed-by: Mika Juoperi <mika.juoperi@nokia.com>