Symbian: Fix Slider tooltip disappears immediately after release
Changed the value indicator tooltip to remain open for a fixed time
after touch release instead of dismissing it immediately. Changed the
pressed property to get unset when keys are released or focus is lost
instead of binding it to the value indicator timer. Changed the tooltip
to follow the position of the handle instead of the mouse to ensure
position is correct also when the handle moves after touch release.

Merge-request: 512
Task-number: QTCOMPONENTS-514
Reviewed-by: Feetu Nyrhinen <ext-feetu.nyrhinen@nokia.com>
Reviewed-by: J-P Nurmi <j-p.nurmi@nokia.com>