Symbian: updated graphic file resolving for button components
ToolButton now uses the updated graphic files. There are now separate
graphics for iconic tool buttons, and in this case the flat property
now defaults to true, because the background graphic is not visible.

Also update TabButton to use segmented graphics when inside ButtonRow.

ToolButton, TabButton, and Button, refactored for latched mode.

Reviewed-by: Mikko Tuokko <mikko.tuokko@nokia.com>
Reviewed-by: Pekka Jokela <pekka.e.jokela@nokia.com>

Merge-request: 691
Reviewed-by: Sami Kananoja <sami.kananoja@nokia.com>