Symbian: TextArea implicit max size binding loop fix
There can be a binding loop if the size of an implicitly sized
TextArea's parent depends on the size of the child. This can be the
case for example with a Column parent that doesn't have both dimensions
of the size explicitly set. The fix is to remove the dependency to the
parent's size from TextArea. The implicit max size is by default not
set (is negative) and the TextArea is allowed to grow infinitely. If
the client sets a value then TextArea is allowed to grow up to it.

Task-number: QTCOMPONENTS-699
Reviewed-by: Taneli Haverinen <ext-taneli.2.haverinen@nokia.com>

Merge-request: 690
Reviewed-by: Samuel Nevala <samuel.nevala@digia.com>