Fixes #272470 - Wrong "common" catalog is loaded when language is set to zh_TW/zh_HK (Traditional chinese)
Details: When found, this patch uses MLocale instead of
QtMobility::QSystemInfo. This is needed, when there are different
catalogs for different territories.

Example: In meego, we have the catalogs common_en.qm and
common_en_US.qm. When the language is set to en_US, the latter catalog
has to be used, when the language is set to en_UK, we have to fall back
to the former one.

Additional note: When no catalog is found, the algorithm just uses
the default. This can be useful in, for example, scratchbox.

Reviewed-By: Sergiy