Fixes: Sheets don't rotate when being direct children of (PageStack)Window
Details: All components inside a (PageStack)Window have to be
rotatable. This is already true for Buttons, Rectangles, Dialogs, ...
but is broken for Sheets. The reason is that Sheets don't check for the
windowContent Item during reparenting.

This patch also ensures that sheets don't get clipped by the status bar
when being direct children of PageStackWindows.

Reviewed-By: Sergiy