Symbian: change ToolBarLayout::backButton so that it's ignored
The backButton property of ToolBarLayout is confusing, its default
value is true, and if you set it to false it doesn't provide any valid
use cases. It's also platform specific. So the plan is to just
deprecate it.

Here its functionality is disabled. Support is added for ButtonRow and
other items to extend into the area of a missing left item.
The meaning of Iconic ToolButton is changed to mean one that doesn't
have the text property set; this means that you can optionally use
ToolButton { visible: false } as a spacer or placeholder in
the ToolBarLayout.

Reviewed-by: Samppa Kaartinen <samppa.kaartinen@nokia.com>

Merge-request: 842
Reviewed-by: Louis Botha <ext-louis.h.botha@nokia.com>