Symbian: Fix Page element losing focus it got due to focus=true
The commit 9c620b677253e16b38bc19c63b65cec7105d50ca added a binding of
the Page's visible-property to a script block that in a normal use
case returned false. This had the side effect that an element of the
Page that used to get the focus because of focus=true didn't get it
anymore. The focus was gained and lost while in createObject. The
problem can be avoided by an alternative solution not binding the
visible-property to a script block. This way the Page's element gets
the focus later and doesn't lose it.

Merge-request: 954
Reviewed-by: Jaakko Korpela <jaakko.korpela@nokia.com>
Reviewed-by: Samuel Nevala <samuel.nevala@digia.com>