Fix feedback when gesture might be a swipe
In the topmost part of the statusbar, press events only arrive once the
compositor has decided that it's not a swipe event. Since in this case,
press- and release event follow each other in short order, the press
feedback is essentially swallowed (enabled, then disabled right away).

This patch makes the visual feedback visible for at least 200ms, so the
behaviour when tapped looks more predictable to the user.