Set fonts according to current language inside of QtComponents.
Details: Adjust the style files in the way that they are using "Nokia
Pure Text" in most languages, but "Arial" when current language is
farsi.

App developers shouldn't use the internal structure deployed with this
patch, but use the UiConstants context property to refer to font
settings.

RevBy: Torsten Rahn