Fixes NB#294095 - China variant: The background cannot be seen
When softwareInputPanel is open and the device is rotated
from portrait to landscape, window is resized to fill the
remaining area of the screen. However, if softwareInputPanel
covers the entire screen, the underlying window height was 0.
Window resizing is no longer done if softwareInputPanel
height equals to the screen height.

Reviewed-by: pending