Symbian: Fix for ButtonRow and ButtonColumn checkedButton
In ButtonRow/ButtonColumn, checkedButton was reset if  visibility
changed. This fix changes the functionality so that even
not visible button can be the checked one. It is application's
responsiblity to handle the situation where checked button is
hidden. Also added tests in componenttest for ButtonRow and
ButtonColumn visibility.

Task-number: https://bugreports.qt.nokia.com/browse/QTCOMPONENTS-1178
Reviewed-by: Mikael Fagerlund <ext-mikael.fagerlund@nokia.com>