Also highlight read-only Text{Field,Area} when selected
The logic here was that read-only Text Items are not highlighted when
selected. They can be selected however, just not accept input. The
latter is indicated by the cursor and the VKB being present, so I don't
think this changed behaviour would be really confusing to the user,
Consistency across apps is probably more important, and this is how it's
done in LMT.

Fixes NB#296882