Don't assume declarative-headers live in QtDeclarative
The QtDeclarative directory is added to the include path when depending
on the module (using QT += declarative), so there's no need to prefix
header includes with QtDeclarative.

The only place where the #include <QtFoo/QFoo.h> form is needed
is in public headers, since doing QT += bar (where bar depends
on foo), will _not_ result in the paths of QtFoo being added to
the include path (only the include paths of QtBar).

Reviewed-by: Sebastian Kügler <sebas@kde.org>