There is no "count" property for QStringListModel and others in C++.
It fixes the cause of following bugreport:
https://bugreports.qt.nokia.com/browse/QTCOMPONENTS-1206

Use length in those cases. Otherwise, the height of the SelectionDialog might
be too small (might actually be invisible entirely to show up any elements).
Actually that is the precise issue in my case, and I need to work it around by
setting some fixed size in QML for instance.

Tested, and works fine on Harmattan.