Add a deselectedIndex property to the SelectionDialog
It would be nice to have an access to the deselectedIndex. A use case that I am
now having is that I would not like to change the Ui at all if the same item was
selected like previously.

It would not hurt any existing users either to have this feature in. It is
currently not possible to nicely make a check whether or not the same item was
selected.