Make the Meego components build against Qt5 with QtQuick2
Most of the differences have been abstracted into a new directory
under src named 'compat', that contains a compatibility layer
between the QtQuick1 APIs and the new QtQuick2 APIs. This reduces
the number of ifdefs that have to be added to the code itself.

A similar approach is used to work around the hard-coding of
QML versions in the QML imports, where qmake is used to pre-
or post-process the sources to bump the version to 2.0

We're not yet building:

 - Examples and demos
 - Unit-tests

The latter is okay for now since most of the tests just verify the API
by setting and getting properties, so if the tests pass using Qt 4.8
they should pass with Qt5 as well (but getting them building requires
adding a bunch of ifdefs all around).