Add icon capability to the ListItem for having drill down and other icon options
This patch makes possible to use icons on the right side of the item for several
purposes, like showing up as an item for SelectionDialog (like the ones in
Settings/Security/Device lock).

It would also allow drillDown icons and so forth, for instance for navigation to
the next page. It would not hurt the existing client codes because the icon
visibility is false by default, and need explicite initialization for becoming
visible.