Cache the display to prevent calling XDamageDestroy with a 0-display
The MDeclarativeStatusBar destructor will destroyXDamageForSharedPixmap
on shutdown, which uses passes the display on to XDamageDestroy. At
that point we no longer have a window or a scene, so there's no way
to get the current display. But, we're not likely to have changed our
display since last time, so we used the cached last known display.