Farsi font fixes for ListDelegate and TumblerButton
Use the correct font when using the Farsi language (which then is Arial,
not Nokia Pure) in the ListDelegates. As it wasn't defined in the
constants used by extra components, add it there as well.

We could go through platformstyle here, but I'd like to avoid creating
another object for every delegate, so we're using the slightly more
verbose version here.