Symbian: Fix InfoBanner text blurriness and icon aspect ratio
The text in InfoBanner component could be blurry when icon was being
used in the banner. The problem appeared when the icon height was an
even number of pixels. In that case, the text top margin got a
non-integer value, which caused the text to be shown blurry in landscape
mode. The problem was most visible when OpenGL paint engine was being
used.

Furthermore, the sourceSize of the icon was set to square
graphicSizeSmall x graphicSizeSmall. This broke the aspect ratio of the
icon if it wasn't square in the source file. Now, only width is set for
the sourceSize which preserves the original aspect ratio.

Task-number: ou1cimx1#1002660

Reviewed-by: Tuija Lindfors <ext-tuija.lindfors@nokia.com>