Fix ScrollDecorator margin handling.
Apparently QML1 allowed evaluating something as either a value or a function,
however, QML2 is a lot less friendly in that regard: only property var can have
functions assigned, and they are not evaluated unless the function is actually
called.

Work around this by always evaluating in a function-type way, and storing them
in property var.