Symbian: Fix orientation problem occuring after camera usage
When orientation aware QQC application was behind a landscape-locked
app, and screen lock was being used, the background app was sometimes
left in wrong orientation once brough to foreground.

This was because, in this particular use case, the background app
receives the layout switch event (from Symbian) already when the
screen lock is lifted, and no new event comes when the foreground app
changes, and landscape locking is removed. This wouldn't matter if
the application would set its layout according the "local" screen
geometry as returned by CAknAppUi::ApplicationRect(), but QQC uses
the "global" orientation setting queried from the window server instead.

The fix is to sync the orientation of the QQC Screen with the system
orientation also when the app gains foreground, in addition to doing that
on layout switch events. Some internal functions were renamed to better
describe their semantics after this change.

Task-number: ou1cimx1#1007193
Reviewed-by: Tuija Lindfors <ext-tuija.lindfors@nokia.com>