Add PageStack::openSheet() and openDialog().
It's not nice to have to do this by hand, as that means lazy loading
is harder. So let's provide a convenience way to do this so we can e.g.
openSheet("MySheet.qml") without having to fuss around.

As the same codepath works for dialogs, add a convenience method for
them too.