Hide toolbar shadows when the toolbar is not visible
Toolbars were made invisible by setting the height to 0, which
doesn't hide the shadows surrounding them. As a result of this and
a bug that creates a toolbar even when none is needed, every Page
without a toolbar had a shadow at the bottom of the screen.

Fixed by setting visible:false after animating the toolbar offscreen.